Must/requirements.txt
flask
langchain
sentence-transformers
faiss-cpu
crewai
streamlit